fastapi/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["name <email>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4"
fastapi = "^0.109"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
prisma = "^0.12"
python-dotenv = "^1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
